chore(docs-lint): address review feedback
Five small improvements following the self-review on #36:

1. docs-lint: switch the freshness check from execSync(\`git log ... "<path>"\`)
   to execFileSync('git', ['log', '-1', '--format=%cI', '--', path]). The
   path is already whitelisted by GITHUB_SOURCE_PREFIX, but using
   execFileSync removes any chance of shell-metacharacter injection via a
   malformed link.

2. docs-lint: improve the section-order error message — now says
   "X appears before Y — reorder so the page reads
   Overview → Method Signature → Examples → Alternatives and Recommendations"
   so a contributor doesn't have to grep the source to see the contract.

3. docs-lint: trim trailing whitespace on heading lines before matching
   the contract, so an accidental "## Overview " doesn't trip the
   required-section check.

4. docs-lint: add an explicit comment in extractGithubLinkPaths
   explaining that tree/ (directory) and blob/<sha>/ (commit-pinned)
   URLs are intentionally skipped — freshness has no meaning for them.

5. docs-link-check: warn on duplicate heading slugs on the same page.
   github-slugger would suffix the second occurrence with "-1", so a
   cross-link targeting the un-suffixed slug silently hits the first
   heading. Now surfaces 4 real collisions in 77.limiters.md (multiple
   "### Configuration" / "### Operation Principle" / "### Features"
   subsections across different limiter components) — warnings, not
   errors, so CI is unaffected.

Also extract the shared markdown walker and frontmatter parser into
scripts/_docs-utils.mjs so the two lint scripts share one
implementation instead of drifting.

All four checks remain green: eslint clean, docs:lint-pages 0/0,
docs:lint-links 0 broken / 4 warnings (existing limiters.md
collisions, see above), typecheck exit 0.
