+++ scripts/docs-link-check.mjs
@@ -1,8 +1,9 @@
 #!/usr/bin/env node
 
-import { readFileSync, readdirSync, statSync } from 'node:fs'
+import { readFileSync } from 'node:fs'
 import { join, resolve, relative, dirname } from 'node:path'
 import { fileURLToPath } from 'node:url'
+import { walkMarkdownFiles, parseFrontmatter, stripFrontmatter } from './_docs-utils.mjs'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const REPO_ROOT = resolve(__dirname, '..')
@@ -10,20 +11,16 @@ const DOCS_ROOT = join(REPO_ROOT, 'docs', 'content', 'docs')
 const URL_PREFIX = '/docs/'
 
 let errors = 0
+let warnings = 0
 
 function logError(file, message) {
   console.log(`\x1B[31mERROR\x1B[0m ${relative(REPO_ROOT, file)}: ${message}`)
   errors++
 }
 
-function walk(dir) {
-  const out = []
-  for (const name of readdirSync(dir)) {
-    const full = join(dir, name)
-    if (statSync(full).isDirectory()) out.push(...walk(full))
-    else if (full.endsWith('.md')) out.push(full)
-  }
-  return out
+function logWarn(file, message) {
+  console.log(`\x1B[33mWARN \x1B[0m ${relative(REPO_ROOT, file)}: ${message}`)
+  warnings++
 }
 
 // Map a file path under DOCS_ROOT to its public URL.
@@ -45,10 +42,9 @@ function fileToUrl(file) {
 
 // Slug a heading the way `github-slugger` (used by Nuxt Content) does:
 // lowercase, strip everything that isn't a word char, whitespace or hyphen,
-// then collapse runs of whitespace into single hyphens.
+// then turn each whitespace char into a single hyphen. Adjacent hyphens are
+// preserved so "A  B" -> "a--b" (matches Nuxt Content's runtime behaviour).
 function slugifyHeading(text) {
-  // Match github-slugger semantics: each whitespace char becomes a single
-  // hyphen (so "A  B" -> "a--b", not "a-b"). Don't collapse adjacent hyphens.
   return text
     .toLowerCase()
     .replace(/[^\w\s-]/g, '')
@@ -56,14 +52,26 @@ function slugifyHeading(text) {
     .replace(/\s/g, '-')
 }
 
-function extractHeadings(body) {
+// Build the slug set and warn on collisions: github-slugger would suffix the
+// second one with `-1`, so a cross-link pointing at the un-suffixed slug would
+// hit the first heading silently. Surfacing the collision lets the author pick
+// distinct titles or accept the explicit `-1` suffix.
+function extractHeadings(file, body) {
   const slugs = new Set()
   for (const line of body.split('\n')) {
     const prefix = line.match(/^#{1,6}\s/)
     if (!prefix) continue
     const text = line.slice(prefix[0].length).trim()
     if (!text) continue
-    slugs.add(slugifyHeading(text))
+    const slug = slugifyHeading(text)
+    if (slugs.has(slug)) {
+      logWarn(
+        file,
+        `heading "${text}" collides with an earlier heading on the same page (both slug to "${slug}"). Internal links to "#${slug}" will hit only the first one.`
+      )
+      continue
+    }
+    slugs.add(slug)
   }
   return slugs
 }
@@ -74,16 +82,9 @@ function buildPageIndex(files) {
   for (const file of files) {
     const raw = readFileSync(file, 'utf8')
     const body = stripFrontmatter(raw)
-    index.set(fileToUrl(file), { headings: extractHeadings(body) })
+    index.set(fileToUrl(file), { headings: extractHeadings(file, body) })
   }
   return index
-}
-
-function stripFrontmatter(text) {
-  if (!text.startsWith('---')) return text
-  const end = text.indexOf('\n---', 3)
-  if (end === -1) return text
-  return text.slice(end + 4).replace(/^\n/, '')
 }
 
 function splitUrlAndFragment(url) {
@@ -107,21 +108,18 @@ function extractInternalLinksFromBody(body) {
   return out
 }
 
-function extractFrontmatter(text) {
-  if (!text.startsWith('---')) return ''
-  const end = text.indexOf('\n---', 3)
-  if (end === -1) return ''
-  return text.slice(3, end)
-}
-
-function extractInternalLinksFromFrontmatter(fm) {
-  // Frontmatter `to:` values that point at /docs/...
+function extractInternalLinksFromLinksArray(linksArray) {
+  // The frontmatter `links:` array is parsed into raw entry strings like
+  // "label: Foo\niconName: GitHubIcon\nto: /docs/foo/". Pull each `to:` line
+  // and keep only entries that target the local doc site.
   const out = []
-  for (const line of fm.split('\n')) {
-    const idx = line.indexOf('to:')
-    if (idx === -1) continue
-    const value = line.slice(idx + 3).trim().replace(/^['"]|['"]$/g, '')
-    if (value.startsWith(URL_PREFIX)) out.push(value)
+  for (const entry of linksArray) {
+    for (const line of entry.split('\n')) {
+      const colon = line.indexOf('to:')
+      if (colon !== 0) continue
+      const value = line.slice(3).trim().replace(/^['"]|['"]$/g, '')
+      if (value.startsWith(URL_PREFIX)) out.push(value)
+    }
   }
   return out
 }
@@ -136,17 +134,16 @@ function resolveLinkTarget(link, sourceUrl) {
 }
 
 function main() {
-  const files = walk(DOCS_ROOT)
+  const files = walkMarkdownFiles(DOCS_ROOT)
   const pageIndex = buildPageIndex(files)
 
   for (const file of files) {
     const sourceUrl = fileToUrl(file)
     const raw = readFileSync(file, 'utf8')
-    const fm = extractFrontmatter(raw)
-    const body = stripFrontmatter(raw)
+    const { frontmatter, body } = parseFrontmatter(raw)
 
     const links = [
-      ...extractInternalLinksFromFrontmatter(fm),
+      ...extractInternalLinksFromLinksArray(frontmatter.links || []),
       ...extractInternalLinksFromBody(body)
     ]
 
@@ -166,7 +163,7 @@ function main() {
     }
   }
 
-  console.log(`\ndocs-link-check: ${errors} broken link(s)`)
+  console.log(`\ndocs-link-check: ${errors} broken link(s), ${warnings} warning(s)`)
   if (errors > 0) process.exit(1)
 }
